requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
Flask_SQLAlchemy==2.5.1
PyJWT==2.6.0
Werkzeug==2.2.3
